Moved quick actions section
