+++ src/app/dashboard/accounts/_components/quick-actions-section.tsx
